requirements.txt
python-telegram-bot
validators